// src/app/models/items.ts
export interface items{
    id_persona: number;
    rut: string;
    nombre: string;
    cargo: string;
    correo: string;
    empresa: string;
    faena: string;
    Vehiculo: string;
    marca: string;
    modelo:string;
    tipo:string;
    tipo_caja_cambio:string;
    tipo_rampla:string;
    Instructor: string;
    fecha: Date;
    estado: string;
    Observacion: string;
    checks:any;
}
export class items{
    id_persona: number;
    rut: string;
    nombre: string;
    cargo: string;
    correo: string;
    empresa: string;
    faena: string;
    Vehiculo: string;
    marca: string;
    modelo:string;
    tipo:string;
    tipo_caja_cambio:string;
    tipo_rampla:string;
    Instructor: string;
    fecha: Date;
    estado: string;
    Observacion: string;
    checks:any}